Update picosock header for int dest_if

+++ picoquic/picosocks.h
@@ -165,13 +165,13 @@ int picoquic_select(SOCKET_TYPE* sockets, int nb_sockets,
 int picoquic_send_through_socket(
     SOCKET_TYPE fd,
     struct sockaddr* addr_dest,
-    struct sockaddr* addr_from, unsigned long from_if,
+    struct sockaddr* addr_from, int from_if,
     const char* bytes, int length, int* sock_err);
 
 int picoquic_send_through_server_sockets(
     picoquic_server_sockets_t* sockets,
     struct sockaddr* addr_dest, 
-    struct sockaddr* addr_from, unsigned long from_if,
+    struct sockaddr* addr_from, int from_if,
     const char* bytes, int length, int * sock_err);
 
 int picoquic_get_server_address(const char* ip_address_text, int server_port,
